requirements.txt
actionlib==1.12.1
angles==1.9.12
autoware_camera_lidar_calibrator==1.12.0
autoware_launcher==1.12.0
bondpy==1.8.5
branca==0.8.1
camera_calibration==1.15.2
camera_calibration_parsers==1.11.13
carla_autoware_bridge==0.0.0
catkin==0.7.29
certifi==2025.4.26
charset-normalizer==3.4.1
contourpy==1.3.2
controller_manager==0.18.4
controller_manager_msgs==0.18.4
cv_bridge==1.13.1
cycler==0.12.1
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
dynamic_reconfigure==1.6.5
folium==0.19.5
fonttools==4.57.0
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
geodesy==0.5.3
gps_common==0.3.4
idna==3.10
image_geometry==1.13.1
interactive-markers==1.11.5
Jinja2==3.1.6
joint_state_publisher==1.12.15
kdl-parser-py==1.13.3
kiwisolver==1.4.8
lanelet2-python==1.0.1
laser_geometry==1.6.7
lgsvl_simulator_bridge==1.12.0
MarkupSafe==3.0.2
matplotlib==3.10.1
message_filters==1.14.13
nmea_navsat_driver==0.5.2
numpy==2.2.5
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.16
python-dateutil==2.9.0.post0
python_qt_binding==0.4.4
pytz==2025.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.32.3
resource_retriever==1.12.7
rosapi==0.11.16
rosbag==1.14.13
rosboost-cfg==1.14.9
rosbridge-library==0.11.16
rosbridge-server==0.11.16
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
roswww==0.1.13
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.3
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rviz==1.13.30
scipy==1.15.2
sensor-msgs==1.12.8
six==1.17.0
smach==2.0.3
smach_ros==2.0.3
smclib==1.8.5
sound-play==0.3.17
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
topic_tools==1.14.13
trafficlight_recognizer==1.12.0
tzdata==2025.2
unique_id==1.0.6
urdfdom-py==0.4.6
urllib3==2.4.0
utm==0.8.1
xacro==1.13.20
xyzservices==2025.4.0
